converting everything to floats basically.

+++ src/Components/EnemyAi.cs
@@ -5,7 +5,7 @@ namespace JustWind.Components
     public class EnemyAi : Component
     {
         public List<Vector2> Path = new();
-        public int Speed = 5;
+        public float Speed = 5;
         public int Level = 1;
         public Vector2 NextTarget;
         public int NextIndex = 0;
@@ -14,7 +14,7 @@ namespace JustWind.Components
         public float MaxScariness = 250;
 
         public double LastTimeDamageDealt = 0;
-        public int Damage = 3;
+        public float Damage = 3;
 
         public EnemyAi()
         {
